Animated swinging pose (Default)
Just made a swinging pose for the unreal default model and a short animation to get into the pose. May/May not work for any unique models because I don't know how modeling works. For now, I'll add this to the character when they grapple and learn how to adjust rotation to aim at grapple location.