Animated grapple effect (COMPLETE)
So, I'm still gonna work on the default character. Already completed a fully implemented animation and all. Now the grapple is automatically connected to the hand and the animation is played when grappling begins. ALL PERFECT. Only things I'd change to make this even better would be to make a grapple animation pose for each of the 8 directions the rope could be in and I already know how to do that as well. (It's just a lot of work is all)