Teleport dagger
The only issue at the moment is that the dagger doesn't ignore collision of the player, making it bounce off the player when thrown into their direction.
I have the layout to set it to ignore collision when overlapping player, and reset itself when leaving collision, but I havent yet implemented it.